# swagger.json
{"openapi":"3.0.0","paths":{"/api/cryptofollow-service/v1":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/cryptofollow-service/v1/investments":{"get":{"operationId":"InvestmentsController_getInvestments","parameters":[],"responses":{"200":{"description":""}},"tags":["investments"]},"post":{"operationId":"InvestmentsController_createInvestment","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInvestmentDto"}}}},"responses":{"201":{"description":""}},"tags":["investments"]}},"/api/cryptofollow-service/v1/investments/{id}":{"get":{"operationId":"InvestmentsController_getInvestment","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["investments"]},"put":{"operationId":"InvestmentsController_updateInvestment","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":1,"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInvestmentDto"}}}},"responses":{"200":{"description":""}},"tags":["investments"]},"delete":{"operationId":"InvestmentsController_deleteInvestment","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["investments"]}},"/api/cryptofollow-service/v1/currency":{"get":{"operationId":"CurrencyController_getCurrencies","parameters":[],"responses":{"200":{"description":""}},"tags":["currency"]},"post":{"operationId":"CurrencyController_createCurrency","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCurrencyDto"}}}},"responses":{"201":{"description":""}},"tags":["currency"]}},"/api/cryptofollow-service/v1/currency/{id}":{"get":{"operationId":"CurrencyController_getCurrencyById","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["currency"]},"put":{"operationId":"CurrencyController_updateCurrency","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":1,"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCurrencyDto"}}}},"responses":{"200":{"description":""}},"tags":["currency"]},"delete":{"operationId":"CurrencyController_deleteCurrency","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["currency"]}},"/api/cryptofollow-service/v1/wallet":{"get":{"operationId":"WalletController_getBalance","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BalanceEntity"}}}}},"tags":["wallet"]}},"/api/cryptofollow-service/v1/wallet/currencies":{"get":{"operationId":"WalletController_getBalanceByCurrencies","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BalanceByCurrencyEntity"}}}}}},"tags":["wallet"]}},"/api/cryptofollow-service/v1/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultErrorResponse"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultErrorResponse"}}}}},"tags":["Users"]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultErrorResponse"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultErrorResponse"}}}}},"tags":["Users"]}},"/api/cryptofollow-service/v1/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultErrorResponse"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultErrorResponse"}}}}},"tags":["Users"]},"put":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":1,"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultErrorResponse"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultErrorResponse"}}}}},"tags":["Users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"5XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultErrorResponse"}}}},"4XX":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultErrorResponse"}}}}},"tags":["Users"]}},"/api/cryptofollow-service/v1/auth/login":{"post":{"operationId":"AuthController_logIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/cryptofollow-service/v1/auth/validate":{"post":{"operationId":"AuthController_validate","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/cryptofollow-service/v1/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}}},"info":{"title":"Crypto Follow Service","description":"Service of Crypto Follow","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateInvestmentDto":{"type":"object","properties":{"buyPrice":{"type":"number","example":0.00003634},"currencyInvestment":{"type":"number","example":2924485},"currencyId":{"type":"number","example":1},"userId":{"type":"number","example":1}},"required":["buyPrice","currencyInvestment","currencyId","userId"]},"CreateCurrencyDto":{"type":"object","properties":{"name":{"type":"string","example":"BONK"},"price":{"type":"number","example":0.00004286},"pair":{"type":"string","example":"USDT"}},"required":["name","price","pair"]},"UpdateCurrencyDto":{"type":"object","properties":{"name":{"type":"string","example":"BONK"},"price":{"type":"number","example":0.00004286},"pair":{"type":"string","example":"USDT"}}},"BalanceEntity":{"type":"object","properties":{"balance":{"type":"number"}},"required":["balance"]},"CurrencyEntity":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string","example":"BONK"},"pair":{"type":"string","example":"USDT"},"price":{"type":"number"}},"required":["id","name","pair","price"]},"BalanceByCurrencyEntity":{"type":"object","properties":{"currencyInvestment":{"type":"number"},"pairInvestment":{"type":"number"},"pairAmount":{"type":"number"},"pairVariation":{"type":"number"},"percentageVariation":{"type":"number"},"currencyId":{"type":"number"},"currency":{"$ref":"#/components/schemas/CurrencyEntity"}},"required":["currencyInvestment","pairInvestment","pairAmount","pairVariation","percentageVariation","currencyId","currency"]},"DefaultErrorResponse":{"type":"object","properties":{"message":{"type":"string","example":"Resource not found"},"error":{"type":"string","example":"Not Found"},"statusCode":{"type":"number","example":404}},"required":["message","error","statusCode"]},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"dev@rozedev.com"},"password":{"type":"string","example":"password"},"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"}},"required":["email","password","firstName","lastName"]},"UserEntity":{"type":"object","properties":{"id":{"type":"number"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"},"loginTries":{"type":"number"},"isEnabled":{"type":"boolean"}},"required":["id","email","firstName","lastName","password","loginTries","isEnabled"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"dev@rozedev.com"},"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"},"loginTries":{"type":"number","example":0},"isEnabled":{"type":"boolean","example":true}},"required":["email","firstName","lastName","loginTries","isEnabled"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]}}}}